requirements.txt
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
requests==2.23.0
Authlib==0.14.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
pip==20.0.2
pycparser==2.20
pyOpenSSL==19.1.0
setuptools==46.0.0
six==1.14.0
urllib3==1.25.8
PyYAML==5.3.1
